Create camera

+++ app/src/main/java/goodlne/info/cardbag/Photo.java
@@ -5,14 +5,14 @@ import android.os.Parcelable;
 
 public class Photo implements Parcelable {
 
-    private int imageID;
+    private long imageID;
 
     public Photo(){
 
         imageID = 0;
     }
 
-    public Photo(int imageID) {
+    public Photo(long imageID) {
 
         this.imageID = imageID;
     }
@@ -34,7 +34,7 @@ public class Photo implements Parcelable {
         }
     };
 
-    public int getImageID() {
+    public long getImageID() {
 
         return imageID;
     }
@@ -46,6 +46,6 @@ public class Photo implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(imageID);
+        dest.writeLong(imageID);
     }
 }
